switch to app factory
