+++ plusplus/__init__.py
@@ -1,0 +1,1 @@
+from plusplus.app import create_app
