+++ plusplus/strings.json
